Don't resolve settings until exec plan is chosen
We can't determine what `:release_path` should be until an execution
plan is built, because the type of plan might be for the "current"
release or for a "new" release.

Therefore defer the interpolation of settings until the time we build
the execution plan. This means that settings can no longer be accessed
from the Runtime object; they can only be accessed from ExecutionPlan.