Fix a bug where bash completions were not working for deploy/run/setup options (#23)
* Restore completions behavior broken by earlier chg

In order to determine completions, `DeployOptions` creates a
hypothetical (empty) execution plan in order to compute the full list of
settings. A recent change added an assertion that prevents empty
execution plans, thus breaking the completions.

Fix by restoring the ability to create empty execution plans.

* Fix an occasional test failure due to race cond.

In some cases when testing locally the test can attempt to open a
connection to the docker container before the sshd process in the
container is ready. Work around this by adding a slight delay after the
starting the docker container.

* DRY up bundler/capture testing utilities