Allow bundler version to specified via setting (#31)
Before, the `bundler:upgrade_bundler` task would work by checking the
`Gemfile.lock` to determine the version of bundler to install. Sometimes
we want to specify the version explicitly instead of relying on this
magic.

In an earlier commit, support for manually specifying a bundler version
was done via a command-line argument, but this is not consistent with
other tomo plugins. Other plugins like nvm and rbenv have a setting that
controls the version. The nice thing about this is settings also work on
the command line via the `-s` option, so it is a nice solution that
supports both styles.

This commit adds a `:bundler_version` setting for this purpose. If the
setting is `nil` (the default), it falls back to the auto-detect
behavior that existed previously. For consistency, the `:run_args`
solution introduced in the earlier commit has been removed.