Fix unknown deploy user on when deploying via Circle CI (#36)
By default Circle CI does not have `USER` or `USERNAME` environment
variables set. We were using these environment variables to determine
the "deploy user" that was written to the revision log and JSON file
during deploy, so as a result this was showing up as "unknown".

Since automating deployments from CI is a common use case, it is worth
working around this in tomo itself (i.e. instead of requiring users to
remember to explicitly set the `USER` environment variable).

Fix by calling `whoami` to get the user name in the case that the
environment variables are missing. This commit places this value in a
setting called `:local_user` so that other tasks can potentially use it
in the future.