Explicitly configure bundler with .bundle/config
Starting with bundler 2.1, flags like `--path` are deprecated for the
`bundle check` and `bundle install` commands. In bundler 3.0 these flags
will be dropped altogether. That means the preferred way to configure
the behavior of bundler going forward is to use either environment
variables or a `.bundle/config` file.

This commit changes how tomo uses bundler to be forward-compatible with
bundler 3.0.  Now, instead of passing flags to `bundle install`, tomo
has a new task called `bundler:config` that will write an appropriate
`.bundle/config` file based on the various `:bundler_*` tomo settings.

This is a breaking change due to the following:

1. The general-purpose `:bundler_install_flags` setting has been
   removed. To configure bundler use one of the more specific settings
   instead.
2. The `.bundle` directory has been removed from `:linked_dirs`. This is
   because tomo no longer relies on the side effects of `bundle install`
   to create the contents of this directory. Instead, tomo now manages
   its own `.bundle` directory (global by default).
3. To configure bundler properly you must add the new `bundler:config`
   task to `setup` just before `bundler:install`.

This commit also adds support for some more bundler options, namely
`retry` and `with`.