Add support for ERB templates (#50)
* Refactor common methods into TaskAPI module

Developers expect a consistent API writing tasks and helpers. Before,
this consistency was implemented through copy-and-paste. Now it is
better enforced via a `TaskAPI` module that is included in both `Remote`
and `TaskLibrary`.

* Add :tomo_config_file_path setting

Automatically set based on where the configuration is loaded from.

* Add template support

Update the `write` helper such that it now optionally accepts a
`template:` keyword argument instead of `text:`. If `template:` is
specified, merge the given template (a local path to an ERB template
file) and use the result and the text to write to the remote location.

For example:

```ruby
remote.write template: File.expand_path("unicorn.service.erb", __dir__),
             to: ".config/systemd/user/unicorn.service"
```

For tasks that wish to deal with templates directly, this commit also
introduces a `merge_template` method that loads an ERB template from a
specified path and merges it.

Both `write template:` and `merge_template` follow this behavior:

- If the template path starts with a `"."`, treat it as relative to the
  tomo config directory
- The ERB template can access `settings`, `paths`, `remote`, etc.

* Document template behavior

* Update custom task tutorial to use merge_template

* More docs

* Pretty error for template-not-found