Manage puma start/stop/restart via systemd (#58)
Before, tomo was using a rudimentary solution for managing puma: it simply
started puma using the `--daemon` flag. This had a few big drawbacks:

1. If the host is rebooted, someone has to remember to manually start puma again
   (i.e. run another `tomo deploy`).
2. Likewise if puma crashes, it will not automatically restart.
3. When puma is restarted via a `tomo deploy`, there is a short period of time
   where puma is not running, and thus HTTP requests get dropped and visitors
   see errors.

This PR reimplements the tomo puma plugin from scratch. The puma plugin now
provides a systemd-based solution for starting, stopping, and restarting puma
using socket activation for zero-downtime restarts. It is based on the best
practices in the official puma documentation.

This implementation installs puma as a _user-level_ service using
`systemctl --user`. This allows puma to be installed, started, stopped, and
restarted without a root user or `sudo`. However, when provisioning the host you
must make sure to run the following command as root to allow the puma process to
continue running even after the tomo deploy user disconnects:

```sh
# run as root
$ loginctl enable-linger <DEPLOY_USER>
```

Stdout and stderr of the puma process will be routed to syslog, as is the
convention for systemd services. For Rails, it is recommended that you set
`RAILS_LOG_TO_STDOUT=1` so that all Rails logs are handled this way (`tomo init`
configures this by default).

The tomo puma plugin assumes that your puma server will listen on a single TCP
port for HTTP (not HTTPS) traffic. In other words, HTTPS termination will be
handled by e.g. Nginx or a separate load balancer.

One challenge of systemd is that it cannot easily be used within a Docker
container. This means that our Docker-based E2E tests cannot use the new puma
plugin out of the box. To work around this, I built simple stubs of the
`systemctl` and `loginctl` systemd tools that can be used during testing. These
mimic just enough systemd behavior to allow an E2E test to pass.

This PR is breaking change. If you have an existing app that uses the puma
plugin, follow these steps to upgrade:

1. Add `puma:setup_systemd` to your `setup` tasks.
2. Add `puma:check_active` to your `deploy` tasks, immediately after
   `puma:restart`.
3. Remove any `puma_*` settings (they have been replaced by new ones; see the
   docs.
4. If you are using a puma port other than the default, 3000, make sure to set
   `:puma_port`.
5. Run `loginctl enable-linger <DEPLOY_USER>` as root on the host.
6. Make sure that puma is stopped on the host.
7. Run `tomo run puma:setup_systemd`.
8. Run `tomo deploy`.

Also check out the updated Deploying Rails from Scratch tutorial.