Change settings interpolation syntax to `%{...}` (#59)
### Problem

Tomo currently uses `%<setting_name>` for string interpolation within settings, 
in order to allow composition of settings like:

```ruby
set application: "my-app"
set deploy_to: "/var/www/%<application>"
```

This works great but it goes against the grain for how `%<...>` is normally used
in Ruby. In Ruby, string formatting (i.e. interpolation) is done in one of three
ways:

```ruby
# All of these produce "/var/www/my-app"
format("/var/www/%s", "my-app")
format("/var/www/%<application>s", application: "my-app")
format("/var/www/%{application}", application: "my-app")
```

Notice that in Ruby's Kernel#format, the angle bracket style requires a type
specifier (e.g. "s"), like this: `%<application>s`. There isn't a precedent for
the tomo style of `%<application>`, i.e. without the "s", and in fact in plain
Ruby it would not produce the desired result.

So if tomo is to be consistent with Ruby conventions, the more appropriate
syntax would be `%{application}`. This style does not use a type.

### Solution

This commit adds `%{...}` as the preferred style for doing settings
interpolation in tomo.

But because suddenly switching from `%<...>` to `%{...}` would be a huge
breaking change to existing users, tomo will continue to support the old
`%<...>` syntax, but with a deprecation warning. In a future version of tomo I
will remove the old syntax entirely.