Send Rails logs to stdout by default
We are using systemd for puma, and systemd has a robust logging system
built in. To take advantage of this, Rails has to be configured to send
its output to stdout. This is something that Rails supports out of the
box via the `RAILS_LOG_TO_STDOUT` environment variable.

Recommend tomo users set this by putting it in the default config file
that is generated by `tomo init`.