Include public/packs in linked_dirs by default (#89)
The official webpacker deployment documentation recommends that
`public/packs` and `node_modudles` be included in `linked_dirs`, so that
webpack compilation does not have be redone from scratch for every
release.

https://github.com/rails/webpacker/blob/master/docs/deployment.md

We already link `node_modules` but `public/packs` was missing. This
commit adds it to the default `linked_dirs` generated by `tomo init` so
that it is included in new tomo projects going forward.