Improve how interactive console is detected
CI is by definition an non-interactive environment. However, before,
tomo was not able to detect this reliably. As a result, if a tomo deploy
used something like `Console.prompt` or `Console.menu`, this would crash
or hang indefinitely.

This commit improves how non-interactive environments are detected so
that tomo exits with a more descriptive error instead of hanging.

- Use env vars to detect whether tomo is running in CI.
- Treat CI environment as non-interactive.
- Console.prompt will now explicitly raise instead of crashing on `nil`.
- Refactor Tomo::Console into a singleton class to facilitate testing.
- Add test coverage.