Improve env:update output and error handling
- Explain why tomo is prompting for a value
- Don't show a giant stack trace if prompting fails due to a non-
  interactive console.

For example, let's say tomo is running a deploy in a CI environment, and
the `env:update` task fails because a required environment variable is
missing and an interactive console is needed.

Before, the user would see a stack trace a message that an interactive
console was needed, but not _why_ it was needed.

Now the error is much more concise and there is an explanation of why
tomo needs an interactive console:

```
$ tomo deploy
tomo deploy v0.17.0
→ Connecting to deployer@app.example.com
• env:update
The remote host needs a value for the $FOO environment variable.
Please provide a value at the prompt.

  ERROR: The env:update task requires an interactive console.

  This appears to be a non-interactive CI environment because the
  CIRCLECI environment variable is set.
```