Fail loudly if tomo cannot guess bundler version (#115)
If `:bundler_version` is not specified, tomo will try to guess the
desired version by extracting the `BUNDLED WITH` value from the
`Gemfile.lock`.

Before, if this fails (e.g. because the lock file is missing), the
`bundler:upgrade_bundler` task would silently do nothing. This is
surprising because no other tomo task behaves like this.

Now, `bundler:upgrade_bundler` will fail with a helpful error message if
the bundler version cannot be determined. If you do not want to use tomo
to manage the installation of bundler, it is easy to just remove the
`bundler:upgrade_bundler` from the list of setup tasks instead of having
it fail silently.