Remove DockerPluginTester
`DockerPluginTester` brought a bunch of extra code, Rakefile complexity,
and CI runtime, but was only being used for a single test. Not worth the
trouble. Remove `DockerPluginTester` and the stuff that was necessary to
support it, including `minitest-hooks`.

To be clear, we still use Docker for the very important e2e test. But
for testing plugins in isolation, Docker is no longer needed.