Use rake to invoke rake tasks, not rails
The documentation for tomo's `rake` helper, and the intent of that
method, is that it runs `bundle exec rake`. However this `rake` helper
was in fact running `bundle exec rails` instead. This was a mistake.

Fix this so that the `rake` helper, and tasks that depend on it (e.g.
`rails:db_migrate`, `rails:db_seed`, etc.) use `bundle exec rake`
instead of `bundle exec rails` as originally intended.

This also makes tomo compatible with older versions of Rails, before the
`rails` was smart enough to be able to invoke rake tasks.