Don't assume "master" as the default git branch
Before, tomo would use "master" as the default value for the
`git_branch` setting if none was explicitly specified in
`.tomo/config.rb`.

The term "master" is problematic given its historical association with
slavery, and many projects are phasing it out in favor of more
appropriate language. GitHub, for example, has said that they will be
using "main" as the default branch name in the future.

To support this, tomo will no longer assume "master" and instead will
have no default value for the `git_branch` setting. Users must
explicitly set it in `.tomo/config.rb`.

To make this easy for new projects, `tomo init` will now automatically
generate a `.tomo/config.rb` file with an appropriate `git_branch`
setting in it. The value of the setting will be determined by inspecting
the current git branch when `tomo init` is called. So if your default
branch is "main", "develop", or whatever, tomo will still work right out
of the box.

This is a breaking change for existing tomo projects where the
`.tomo/config.rb` file had no `git_branch` explicitly specified; i.e.
for projects relying on the default "master" value. If you have a
project like this, simply add this line when upgrading:

```ruby
set git_branch: "master"
```