Auto-disable Style/FormatStringToken rubocop rule
Tomo relies on `%{...}` syntax for interpolating settings. However, the
default configuration for rubocop flags this syntax as a problem and
suggests an alternative that doesn't work with tomo.

To prevent confusion, `tomo init` will now automatically include a
special comment to disable the `Style/FormatStringToken` rule if it
detects that rubocop is present.