Prevent multiple apps from sharing the same bashrc (#155)
There is a fundamental assumption in the tomo "env" tasks that only one
application is being deployed to a given user@host. Some users have
reported being confused and surprised when they deploy more than one app
to the same user@host and things begin behaving strangely.

This limitation is documented in tomo's FAQ, but it would be even better
if tomo would alert users earlier in the setup process before the second
app is deployed and things start going sideways.

This commit modifies the `env:setup` task so that it is able to detect
when a second app is being deployed to the same user@host. When this
situation is detected it will halt the setup process with a detailed
error message like this:

```
tomo run v1.2.0
→ Connecting to deployer@app.example.com
• env:setup
cat .bashrc

  ERROR: The env:setup task failed on deployer@app.example.com.

  Based on the contents of .bashrc, it looks like another application
  is already being deployed via tomo to this host, using the following envrc
  path:

    /home/deployer/apps/example/envrc

  Tomo is designed such that only one application can be deployed to a given
  user@host. To deploy multiple applications to the same host, use a separate
  deployer user per app. Refer to the tomo FAQ for details:

    https://tomo-deploy.com/#faq

  You may be receiving this message in error if you recently renamed or
  reconfigured your application. In this case, remove the references to the
  old envrc path in the host's .bashrc and re-run env:setup.
```

Fixes #137, #138.