Generate config.rb with explicit Ruby version, if needed
Problem:

If a project lacks a `.ruby-version` file, then `tomo setup` will fail
if `rbenv_ruby_version` is not explicitly specified. Currently tomo does
not include a `rbenv_ruby_version` setting in the default
`.tomo/config.rb` file.

Solution

We can make the user's life a lot easier in this scenario by providing a
default value when they initialize tomo.

This commit enhances `tomo init` so that if a `.ruby-version` is not
present, tomo will include a `set rbenv_ruby_version: ...` setting in
the generated `.tomo/config.rb` based on the currently executing version
of Ruby.