Remove default bundler_jobs value of "4" (#147)
Bundler will automatically determine the "best" value to use for the
jobs options based on the number of processors, so there is no need to
specify a value. Especially an arbitrary value of "4" that may not make
sense depending on the machine where bundler is being used.

This commit removes tomo's default value of the `bundler_jobs` setting
so that Bundler's built-in behavior will be used instead. The
`bundler_jobs` setting is still supported if tomo user's want to provide
a specific value.