Gracefully handle non-standard .ruby-version files
`tomo init` looks for the presence of a `.ruby-version` file. If that
file exists, then the `.tomo/config.rb` file will omit an explicit
setting for `:rbenv_ruby_version`. That's because tomo assumes it will
be able to auto-detect the version at setup time.

However, if `.ruby-version` contains bad data, or a value that rbenv
doesn't understand, that means that the user will be in for an
unpleasant surprise: `tomo setup` will fail because it tries to guess
the `:rbenv_ruby_version` from the `.ruby-version` file, and fails.

This commit fixes this problem with the following solution:

`tomo init` will omit the `:rbenv_ruby_version` setting only if the
`.ruby-version` file contains an actual Ruby version number and that
version number matches `RUBY_VERSION`. That way the auto-detection
should be guaranteed to work.

Fixes #211