Suggest `bundle exec` to fix missing plugin errors
If a user installs tomo and tomo plugins via a Gemfile, but then runs
tomo without using `bundle exec`, this can lead to errors like this:

```
ERROR: sidekiq is not a recognized plugin.
You may need to install the tomo-plugin-sidekiq gem.
```

This is because the plugins in the Gemfile are not available when tomo
is run without `bundle exec`.

Until now, tomo's error message doesn't really make this clear.

This commit improves the error message so that it now says:

```
ERROR: sidekiq is not a recognized plugin.
You may need to install the tomo-plugin-sidekiq gem.
Try prefixing the tomo command with bundle exec to fix this error.
```

tomo will only suggest `bundle exec` if it is fairly certain that it
will fix the error: there is a Gemfile present and it contains a `gem`
entry for the missing plugin.