Allow yarn to be installed globally w/o version
Starting with yarn v2, the global version of yarn is 1.x, and it is up
to the project to provide its desired version of yarn.

For tomo, that introduces some ambiguity. The nodenv plugin has a
setting for `:nodenv_yarn_version`. Which version is this talking about?
The global version of yarn or the project one?

To address this confusion, this commit changes tomo's behavior to no
longer require `:nodenv_yarn_version`. Instead, tomo will always install
yarn globally by default, without a version specifier. This will result
in yarn 1.x getting installed globally, which is what is appropriate for
typical yarn usage.

If a project needs a specific version of yarn, the proper way to do that
is within the source code of the project itself (`.yarnrc.yml`), not by
installing a different global version of yarn.

The `:nodenv_node_version` setting remains in tomo, but it is no longer
recommended.