Ensure git is configured with a valid `user.name` and `user.email` (#280)
Certain operations, like the `git pull` issued by the `rbenv` installer,
require that the git client on the remote host is configured with a
valid `user.name` and `user.email`.

Git will try to auto-detect values for these settings, but depending on
the host, this may not always be possible.

When this auto-detection fails, tomo will fail with errors like this:

```
*** Please tell me who you are.

Run

  git config --global user.email "you@example.com"
  git config --global user.name "Your Name"

to set your account's default identity.
Omit --global to set the identity only in this repository.

fatal: unable to auto-detect email address
```

Fix by introducing a new tomo task that runs during setup: `git:config`.
This will explicitly set the `user.name` and `user.email` values, so
that auto-detection is no longer necessary.