Fix "uninitialized constant" errors for StringIO and SimpleDelegator
In instances when tomo is used without a Gemfile, or is used with a
Gemfile that contains very few gems, running `tomo` might result in an
"uninitialized constant" crash for `StringIO` or `SimpleDelegator`.

This happens because tomo is trying to use these constants without
explicitly `require`-ing them. By luck, gems in a typical Gemfile will
already require these, so by the time tomo runs, they have already been
loaded. But this is not guaranteed.

Fix by using the proper `require` statements to insure that these
constants are loaded.