Fix bug where `nodenv:install` would fail during `--dry-run` (#373)
When `nodenv:install` tries to guess what Node version to use, it
attempts to read a `.node-version` file. This doesn't work in a
`--dry-run`, because no SSH I/O occurs and the file is not actually
read. As a result, a dry run would lead to this error:

> Could not guess node version from .node-version file.
Use the :nodenv_node_version setting to specify the version of node to
install.

Fix by using a placeholder node version in the dry run scenario.

This PR introduces tomo's first dry-run unit test, so a little extra
test support code was also added to make this type of test possible.