Configure puma.service with Type=notify by default (#381)
### New behavior

As of late 2020, the official Puma docs suggest configuring the Puma
systemd service with `Type=notify`. This has been supported with CRuby
since Puma 5.1. The `notify` option allows for Puma to be automatically
restarted if it locks up.

https://github.com/puma/puma/blob/master/docs/systemd.md

This PR changes tomo to use `Type=notify` when configuring Puma going
forward.

### Opting out

For tomo projects that are on JRuby or are using Puma < 5.1, the
previous behavior can be restored with the following tomo setting:

```ruby
set puma_systemd_service_type: "simple"
```