Rails deploy tutorial: clarify that force_ssl is disabled (#390)
The tomo Rails tutorial is simple in part because it excludes SSL. This
has worked fine with the behavior of `rails new`, which has SSL disabled
by default.

However, starting with the forthcoming Rails 7.1, `rails new` will
generate a production configuration that enables SSL by default. This
will interfere with the simplicity of the tomo deploy tutorial.

This commit updates the tomo documentation to clarify that `force_ssl`
needs to be set to `false` for the purposes of the Rails deploy
tutorial.