More reliably guess the right git_branch when generating .tomo/config.rb
This PR improves how `tomo init` fills in the default `git_branch` value
in the generated `.tomo/config.rb` file.

**Improved behavior on new, empty git repos**

Before, when `tomo init` was run on a brand new git repo that does not
yet have any commits, a confusing message would be printed to stderr and
tomo would fail to fill in the `git_branch` setting in the generated
`.tomo/config.rb` file:

```
fatal: ambiguous argument 'HEAD': unknown revision or path not in the working tree.
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'
```

Now, `tomo init` will use the `git branch --show-current` command
(available in git 2.22 and newer) to determine the current branch name.
This command works even for a git repo that does not yet have any
commits. In this case git returns the default branch name. For example,
this will usually resolve to `main`.

**Improved behavior on feature branches**

When adding tomo to an existing project, developers will often be
working on a feature branch, not their main branch.

Before, `tomo init` would always use the current branch name when
filling in the default `git_branch` setting. This is not usually what
you want, because feature branches are short-lived.

Now, `tomo init` will try to guess the main branch name and use that
instead. It checks if either `main` or `master` exists, and if so, it
will use that for the default `git_branch` setting. Otherwise, it will
fall back to the existing behavior of using the current branch name.