Fix `TypeError` when `tomo/config.rb` uses frozen string literals (#448)
If the `.tomo/config.rb` file declares a setup or deploy task with
`privileged: true` and the file starts with the special `#
frozen_string_literal: true` comment, tomo crashes with the following
error message:

```
TypeError: can't define singleton
    lib/tomo/configuration/dsl/tasks_block.rb:17:in `extend_object'
    lib/tomo/configuration/dsl/tasks_block.rb:17:in `extend'
    lib/tomo/configuration/dsl/tasks_block.rb:17:in `run'
```

This is because tomo is trying to call `.extend` on a frozen string.

Fix by creating a mutable copy of task names before extending them.