Fix `FrozenError` when using interactive environment selector (#452)
When running `tomo deploy` in a multi-environment project with no
environment explicitly specified, tomo prompts:

```
Choose an environment: 
❯ staging
  production
```

Interacting with this menu with the keyboard would result in an error:

```
forwardable.rb:240:in 'String#clear': can't modify frozen String: "" (FrozenError)
	from forwardable.rb:240:in 'Tomo::Console::KeyReader#clear'
```

This is because the internal `buffer` used by `KeyReader` was frozen as
a side effect of #446.

Fix by marking `buffer` as mutable.